Change the result of nbMarking from int to double

+++ Sources/SVSKit/SVS.swift
@@ -214,13 +214,13 @@ public struct SVS {
     return SVS(values: res)
   }
   
-  public func nbOfMarkings() -> Int {
+  public func nbOfMarkings() -> Double {
     if self.isEmpty {
       return 0
     }
     let net = self.first!.net
     let markingCapacity = Marking(net.capacity, net: net)
-    var res: Int = 0
+    var res: Double = 0
     for sv in self.values {
       res += sv.nbOfMarkings()
     }
